+++ src/components/marketingStatusLabel/__tests__/MarketingStatusLabel.spec.tsx
@@ -7,4 +7,12 @@ describe('Component - MarketingStatusLabel', () => {
     const { asFragment } = render(<MarketingStatusLabel>Upgrade now</MarketingStatusLabel>);
     expect(asFragment()).toMatchSnapshot();
   });
+
+  it('renders with an icon', async () => {
+    const Icon = <span>Fake icon</span>;
+    const screen = render(<MarketingStatusLabel icon={Icon}>Upgrade now</MarketingStatusLabel>);
+
+    expect(screen.getByText('Fake icon')).toBeVisible();
+    expect(screen.asFragment()).toMatchSnapshot();
+  });
 });